+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.9.3
+* Delete `Geometry::device_idx_for_trans_idx`, use `Geometry::global_to_local_idx` instead
+    * Delete `AUTDDeviceIdxForTransIdx` from capi
+
 # 1.9.2
 * Fix #10
 * Delete `Geometry::del_device`, `Geometry::clear_devices`, `Geometry::direction`, and `Geometry::group_id_for_device_idx`